+++ src/Symfony/Bundle/FrameworkBundle/CHANGELOG.md
@@ -19,6 +19,7 @@ CHANGELOG
  * Added support for Translator paths, Twig paths in translation commands.
  * Added support for PHP files with translations in translation commands.
  * Added support for boolean container parameters within routes.
+ * Added the `messenger:setup-transports` command to setup messenger transports
 
 4.2.0
 -----